No login possible without email confirmation
The registered user is required to click the link send to it's email in
order to confirm it's email before it can access it's account.

It is necessary to have a confirmed email in case that the user does not
remember its password, because a reset link will be sent to that email.